Test kubevirt_ssp_vm_rbd_block_volume_without_rxbounce
Test for metric kubevirt_ssp_vm_rbd_block_volume_without_rxbounce.
